+++ test/utils.js
@@ -2,7 +2,7 @@ import nock from 'nock';
 
 module.exports = {
   createNock() {
-    return nock('http://ps.pubnub.com:80', {
+    return nock('http://ps.pndsn.com:80', {
       filteringScope: () => true
     });
   }